requirements.txt
psycopg2-binary==2.9.6
pandas==1.5.3
matplotlib==3.7.1
plotly==5.22.0
openpyxl==3.1.2